pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.0"
dependencies = [
    "celery>=5.5.2",
    "fastapi[standard]>=0.112.2",
    "pydantic-settings>=2.9.1",
    "pymysql>=1.1.1",
    "pytz>=2025.2",
    "redis>=6.1.0",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.41",
    "flower>=2.0.1",
    "openai>=1.92.2",
    "pymupdf>=1.26.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
hello = "uv_docker_example:hello"

[tool.hatch.build.targets.wheel]
packages = ["src/api"]

[tool.uv]
index-url = "https://mirrors.aliyun.com/pypi/simple/"

[dependency-groups]
dev = [
    "fastapi-cli>=0.0.7",
    "pytest>=8.3.5",
]

